[bugfix] Fixed major bug in InfiniTAM's hashing scheme
The original implementation had live frame allocated in a single pass,
whereas the hashIndex method produced the same hash values for different
blocks (i.e. -6,-6, 90 and 6, 6, 90 both yield hash 122790). This resulted
in coordinates of blocks-to-be-allocated to be overwritten, which
in turn resulted in hash blocks missing from the scene.

The bugfix is a temporary patch. It is unclear how to deal with the
collision prevention in a thread-safe manner: the patch results in
potential data-races. Ideally, atomic compare-and-swap should be used.