[bugfix] scene traversal moved to a templated class
Apparently, linker is insensitive to identical templated function signatures
in separate classes, so the code an a particular file can actually end
up using a function which is not in the included header hierarchy of the
said file. In order to prevent this
(what seems to be - hold your breath - compiler error!), I had to
hide away functions with matching signatures inside a specialized templated class.